Move list item declarations
