+++ src/dashboard/layout/SecondaryList.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="text-xs font-semibold leading-6 text-gray-400">{{ title }}</div>
+  <div class="text-gray-400 text-xs font-semibold leading-6">{{ title }}</div>
   <ul
     role="list"
     class="-mx-2 mt-2 space-y-1"
@@ -24,7 +24,7 @@
         />
         <span
           v-else
-          class="flex h-6 w-6 shrink-0 items-center justify-center rounded-lg border border-gray-700 bg-gray-800 text-[0.625rem] font-medium text-gray-400 group-hover:text-white"
+          class="border-gray-700 bg-gray-800 text-gray-400 group-hover:text-white flex h-6 w-6 shrink-0 items-center justify-center rounded-lg border text-[0.625rem] font-medium"
         >
           {{ item.initial ?? '?' }}
         </span>
@@ -34,47 +34,17 @@
   </ul>
 </template>
 
-<script setup lang="ts">
-import { useDashboardSidebarStore } from '@/dashboard/stores/dashboardSidebar';
-import type { FunctionalComponent } from 'vue';
-
-interface ListItemBase {
-  /** The visible name of the item */
-  name: string;
-  /** An optional id to use as the element key - otherwise name. */
-  id?: string | number;
-  /** An optional value to trace the active state, used to change visual */
-  current?: boolean;
-}
-interface ListItem_Icon extends ListItemBase {
-  icon: FunctionalComponent;
-}
-interface ListItem_Initial extends ListItemBase {
-  initial: string;
-}
-
-interface ListItem_Href {
-  href: string;
-}
-interface ListItem_Click {
-  click: () => void | Promise<void>;
-}
-
-type LinkListItem =
-  | (ListItem_Href & ListItemBase)
-  | (ListItem_Href & ListItem_Initial)
-  | (ListItem_Href & ListItem_Icon);
-type ClickListItem =
-  | (ListItem_Click & ListItemBase)
-  | (ListItem_Click & ListItem_Initial)
-  | (ListItem_Click & ListItem_Icon);
-
-export type ListItem = LinkListItem | ClickListItem;
+<script lang="ts">
+import type { ListItemBase } from './ListItems';
 
 export interface SecondaryListProps {
   title?: string;
   items?: ListItemBase[];
 }
+</script>
+
+<script setup lang="ts">
+import { useDashboardSidebarStore } from '@/dashboard/stores/dashboardSidebar';
 
 const props = defineProps<SecondaryListProps>();
 const sidebarStore = useDashboardSidebarStore();
